chore: rename component-list to components
